Refactor LTX one-stage entry points
