+++ tests/diffusion/models/ltx2/test_ltx2_3_image2video_pipeline.py
@@ -21,7 +21,7 @@ def _make_ltx23_request_pipe(cls):
 
 class TestLTX23ImageToVideoForwardStages:
     def test_forward_resolves_request_image_and_delegates_to_shared_forward_impl(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
         from vllm_omni.diffusion.request import OmniDiffusionRequest
         from vllm_omni.diffusion.worker.request_batch import DiffusionRequestBatch
         from vllm_omni.inputs.data import OmniDiffusionSamplingParams
@@ -66,7 +66,7 @@ class TestLTX23ImageToVideoForwardStages:
         assert seen["kwargs"]["noise_scale"] == 0.5
 
     def test_denoise_timestep_kwargs_masks_video_only(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         ts = torch.tensor([2.0, 4.0])
@@ -88,14 +88,13 @@ class TestLTX23ImageToVideoForwardStages:
 
 class TestLTX23ImageToVideoPipeline:
     def test_ltx23_i2v_pipeline_reuses_ltx23_semantics(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23Pipeline
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline, LTX23Pipeline
 
         assert issubclass(LTX23ImageToVideoPipeline, LTX23Pipeline)
         assert LTX23ImageToVideoPipeline.support_image_input is True
 
     def test_ltx23_i2v_rejects_multi_image_prompt_list(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         image = object()
 
@@ -104,7 +103,7 @@ class TestLTX23ImageToVideoPipeline:
             LTX23ImageToVideoPipeline._resolve_single_prompt_image([object(), object()])
 
     def test_ltx23_i2v_additional_image_resolution_is_tensor_safe(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         image = torch.zeros(1, 3, 4, 4)
         additional = {
@@ -118,7 +117,7 @@ class TestLTX23ImageToVideoPipeline:
     def test_ltx23_i2v_packed_latents_are_not_noised(self, monkeypatch):
         import vllm_omni.diffusion.models.ltx2.ltx2_conditioning as ltx2_conditioning
         import vllm_omni.diffusion.models.ltx2.ltx2_latents as ltx2_latents
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         torch.nn.Module.__init__(pipe)
@@ -154,7 +153,7 @@ class TestLTX23ImageToVideoPipeline:
     def test_ltx23_i2v_5d_latents_noise_preserves_conditioning_frame(self, monkeypatch):
         import vllm_omni.diffusion.models.ltx2.ltx2_conditioning as ltx2_conditioning
         import vllm_omni.diffusion.models.ltx2.ltx2_latents as ltx2_latents
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         torch.nn.Module.__init__(pipe)
@@ -193,7 +192,7 @@ class TestLTX23ImageToVideoPipeline:
         torch.testing.assert_close(out, torch.tensor([[[10.0, 11.0], [1.0, 1.0], [1.0, 1.0]]]))
 
     def test_ltx23_i2v_video_step_preserves_conditioning_frame(self):
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         pipe = object.__new__(LTX23ImageToVideoPipeline)
         torch.nn.Module.__init__(pipe)
@@ -222,24 +221,24 @@ class TestLTX23ImageToVideoPipeline:
 
 
 class TestLTX23ImageToVideoModule:
-    """Test that pipeline_ltx2_3_image2video.py exposes I2V entry points."""
+    """Test that the shared one-stage module exposes LTX2.3 I2V entry points."""
 
     def test_i2v_classes_importable(self):
         """I2V classes must be importable from the implementation module."""
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import LTX23ImageToVideoPipeline
 
         assert LTX23ImageToVideoPipeline is not None
 
     def test_post_process_func_importable(self):
         """get_ltx2_post_process_func must be importable for registry lookup."""
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import get_ltx2_post_process_func
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import get_ltx2_post_process_func
 
         assert callable(get_ltx2_post_process_func)
 
     def test_i2v_class_matches_package_export(self):
         """Package export must point at the I2V implementation module."""
         from vllm_omni.diffusion.models.ltx2 import LTX23ImageToVideoPipeline as PackageExported
-        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import (
+        from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import (
             LTX23ImageToVideoPipeline as ModuleExported,
         )
