+++ vllm_omni/diffusion/models/ltx2/__init__.py
@@ -7,19 +7,19 @@ from vllm_omni.diffusion.models.ltx2.ltx2_components import (
 )
 from vllm_omni.diffusion.models.ltx2.ltx2_transformer import LTX2VideoTransformer3DModel
 from vllm_omni.diffusion.models.ltx2.pipeline_ltx2 import (
-    LTX2Pipeline,
-    LTX2T2VDMD2Pipeline,
-    LTX2TwoStagesPipeline,
-    get_ltx2_post_process_func,
-)
-from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3 import LTX23Pipeline
-from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_3_image2video import LTX23ImageToVideoPipeline
-from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_image2video import (
     LTX2I2VDMD2Pipeline,
     LTX2ImageToVideoPipeline,
-    LTX2ImageToVideoTwoStagesPipeline,
+    LTX2Pipeline,
+    LTX2T2VDMD2Pipeline,
+    LTX23ImageToVideoPipeline,
+    LTX23Pipeline,
+    get_ltx2_post_process_func,
 )
 from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_latent_upsample import LTX2LatentUpsamplePipeline
+from vllm_omni.diffusion.models.ltx2.pipeline_ltx2_two_stage import (
+    LTX2ImageToVideoTwoStagesPipeline,
+    LTX2TwoStagesPipeline,
+)
 
 __all__ = [
     "LTX2Pipeline",
